test(e2e): extract the shared W2 fixture server into uxfix_fixture.py
The slice-1 and slice-2 rigs each carried a verbatim copy of the deterministic
W2 messy-reality fixture server (§4.2) — the duplication the slice-2 verifier
flagged. One module now owns the frozen-NOW0 dataset, the same-origin build
step, the /api/v1 + /ws handler, and the POST /__fixture switches; both rigs
import it, with their assertions untouched. Verified: both rigs green against
the current main build after the refactor.
