+++ e2e/uxfix_fixture.py
@@ -1,0 +1,295 @@
+#!/usr/bin/env python3
+"""
+uxfix_fixture.py — the ONE deterministic W2 messy-reality fixture server
+(DES-UXFIX-001 §4.2) shared by every uxfix slice rig.
+
+Extracted from uxfix_slice1_test.py / uxfix_slice2_test.py (which previously
+each carried a verbatim copy, as the slice-2 verifier flagged): a
+ThreadingHTTPServer that serves the `dist-sameorigin/` build (the no-
+VITE_API_HOST build — `apiBase()` derives from window.location, so the page,
+the API and the `/ws` handshake share ONE origin, no rebuild) plus every
+endpoint the home route reads, with all timestamps computed from a single NOW0
+captured at import. No crew daemon is involved anywhere.
+
+The dataset is §4.2's rows verbatim, including the two adjacencies the rigs
+must not lose:
+  - `legacy-spike`: run failed 8 DAYS ago but its project was touched an HOUR
+    ago — the R3 trap the `runEvents` backfill exists to defuse;
+  - `upload-endpoint`: live, narrating over the rig's own /ws.
+
+Mutable switches (flipped over POST /__fixture between page loads):
+  orphan          — whether the orphan run rides the run list (default True)
+  q3_gate_age_ms  — the r-q3 gate's receivedAt age (default 30s)
+
+A rig that never flips them gets the default W2 board.
+"""
+
+import base64
+import hashlib
+import json
+import os
+import subprocess
+import threading
+import time
+import urllib.parse
+import urllib.request
+from functools import partial
+from http.server import SimpleHTTPRequestHandler, ThreadingHTTPServer
+from pathlib import Path
+
+REPO = Path(__file__).resolve().parent.parent
+SHOTS = REPO / "e2e" / "shots" / "uxfix"
+NPM = "npm.cmd" if os.name == "nt" else "npm"
+
+# Same rule as the main harness: gate toasts are not these surfaces (and the
+# home route does not render them), but the suppression is cheap and display-only.
+HIDE_GATE_TOASTS = '[data-testid="gate-notification"] { display: none !important; }'
+
+# ── The frozen clock (§4.0 determinism): every age derives from this one NOW0 ──
+NOW0 = int(time.time() * 1000)
+SEC, MIN, HOUR, DAY = 1_000, 60_000, 3_600_000, 86_400_000
+
+
+def iso(ms: int) -> str:
+    return time.strftime("%Y-%m-%dT%H:%M:%S", time.gmtime(ms / 1000)) + f".{ms % 1000:03d}Z"
+
+
+# Mutable fixture switches, flipped over POST /__fixture between page loads.
+state = {"orphan": True, "q3_gate_age_ms": 30 * SEC}
+state_lock = threading.Lock()
+
+
+def project(pid: str, name: str, updated_at: int, **extra) -> dict:
+    return {"id": pid, "name": name, "description": None, "status": "active",
+            "scope": f"project:{pid}" if pid != "default" else "",
+            "created_at": updated_at, "updated_at": updated_at, **extra}
+
+
+# §4.2's rows. `legacy-spike`'s project was touched an HOUR ago while its run
+# failed 8 DAYS ago — the exact R3 trap the runEvents backfill exists to defuse.
+QUIET_CLONES = [project(f"quiet-{i:02d}", f"quiet-clone-{i:02d}", NOW0 - 3 * DAY - i * 7 * HOUR)
+                for i in range(20)]
+PROJECTS = [
+    project("default", "Unfiled", NOW0),  # synthesized — must never render (F5)
+    project("legacy-spike", "legacy-spike", NOW0 - HOUR),
+    project("upload-endpoint", "upload-endpoint", NOW0),
+    project("q3-review-deck", "q3-review-deck", NOW0 - 30 * SEC),
+    project("api-migration", "api-migration", NOW0 - 2 * MIN),
+    project("auth-refactor", "auth-refactor", NOW0 - 12 * MIN),
+    project("smoke-tests", "smoke-tests", NOW0 - 6 * DAY),
+    project("notes", "notes", NOW0 - 2 * DAY, interactiveRoot="/tmp/wi-notes"),
+    project("scratch", "scratch", NOW0),
+] + QUIET_CLONES
+
+MEMBERS = {
+    "legacy-spike": ["r-legacy"],
+    "upload-endpoint": ["r-upload"],
+    "q3-review-deck": ["r-q3"],
+    "api-migration": ["r-api"],
+    "auth-refactor": ["r-auth"],
+    "smoke-tests": ["r-smoke1", "r-smoke2"],
+}
+
+
+def session(rid: str, status: str, problem: str, unit_desc: str) -> dict:
+    return {"session": {
+        "id": rid, "workflow_id": "wf-w2", "problem": problem, "entity_mode": "shared",
+        "collection_scope": None, "clis": ["claude"], "status": status,
+        "human_confirm": "all" if status == "awaiting_human" else "none",
+        "unit_ix": 0, "attempt": 0, "workdir": None, "repo_ref": None,
+        "extra_write_roots": [], "archived_at": None, "archive_note": None,
+    }, "units": [{
+        "id": f"{rid}:u0", "session_id": rid, "ord": 0, "description": unit_desc,
+        "stage": "build", "assigned_cli": None, "assigned_invocation": None,
+        "council_task_ref": None, "routing": None, "denial_reason": None,
+        "phase_ref": None, "conformance_ref": None, "phase_status": None,
+        "collection_scope": None, "status": "pending",
+    }]}
+
+
+RUNS = [
+    session("r-q3", "awaiting_human", "make the Q3 review deck", "author the deck outline"),
+    session("r-api", "awaiting_human", "migrate the auth tables", "plan the migration"),
+    session("r-upload", "executing", "add rate-limiting to the upload endpoint",
+            "add rate-limiting to the upload endpoint"),
+    session("r-auth", "failed", "refactor the auth middleware", "refactor the auth middleware"),
+    session("r-legacy", "failed", "spike the legacy importer", "spike the legacy importer"),
+    session("r-smoke1", "completed", "smoke: login flow", "smoke the login flow"),
+    session("r-smoke2", "completed", "smoke: checkout flow", "smoke the checkout flow"),
+]
+ORPHAN = session("r-orphan", "executing", "stranded work from another client",
+                 "stranded work from another client")
+
+# The durable-log tails (D3 step 2): the ONE honest clock for a failure's age.
+RUN_EVENTS = {
+    "r-legacy": [{"type": "sessionStarted", "session": "r-legacy", "ts": NOW0 - 8 * DAY - MIN},
+                 {"type": "sessionFailed", "session": "r-legacy", "ts": NOW0 - 8 * DAY}],
+    "r-auth": [{"type": "sessionStarted", "session": "r-auth", "ts": NOW0 - 13 * MIN},
+               {"type": "sessionFailed", "session": "r-auth", "ts": NOW0 - 12 * MIN}],
+}
+
+NOTES_DOCS = [
+    {"name": "ideas", "kind": "doc", "head": 1, "versions": 1, "updated_at": iso(NOW0 - 2 * DAY)},
+    {"name": "todo", "kind": "doc", "head": 1, "versions": 1, "updated_at": iso(NOW0 - 3 * DAY)},
+]
+
+WS_GUID = "258EAFA5-E914-47DA-95CA-C5AB0DC85B11"
+NARRATION = "Writing the token-bucket middleware for /upload"
+
+
+def ws_frame(payload: dict) -> bytes:
+    data = json.dumps(payload).encode()
+    if len(data) < 126:
+        head = bytes([0x81, len(data)])
+    else:
+        head = bytes([0x81, 126]) + len(data).to_bytes(2, "big")
+    return head + data
+
+
+class W2Handler(SimpleHTTPRequestHandler):
+    """SPA + the whole /api/v1 surface the home route reads + /ws, one origin."""
+
+    def log_message(self, *_args):  # keep stdout JSON-clean
+        pass
+
+    def _json(self, status: int, payload) -> None:
+        body = json.dumps(payload).encode()
+        self.send_response(status)
+        self.send_header("Content-Type", "application/json")
+        self.send_header("Content-Length", str(len(body)))
+        self.end_headers()
+        self.wfile.write(body)
+
+    def _ws(self) -> None:
+        """Accept the upgrade, then stream `unitOutputDelta` frames for the live
+        run — `useRuns` gates its first fetch on a connected socket, so the
+        handshake is mandatory, and the narration keeps the headline honest."""
+        key = self.headers.get("Sec-WebSocket-Key", "")
+        accept = base64.b64encode(hashlib.sha1((key + WS_GUID).encode()).digest()).decode()
+        self.wfile.write(
+            ("HTTP/1.1 101 Switching Protocols\r\nUpgrade: websocket\r\n"
+             f"Connection: Upgrade\r\nSec-WebSocket-Accept: {accept}\r\n\r\n").encode())
+        self.wfile.flush()
+        try:
+            while True:
+                self.wfile.write(ws_frame({
+                    "type": "unitOutputDelta", "session": "r-upload", "ord": 0,
+                    "text": NARRATION + "\n",
+                }))
+                self.wfile.flush()
+                time.sleep(1.0)
+        except OSError:
+            pass
+        self.close_connection = True
+
+    def _api(self, path: str) -> bool:
+        if path == "/api/v1/health":
+            self._json(200, {"status": "ok", "version": "w2-fixture", "ping": "pong"})
+            return True
+        if path == "/api/v1/runs":
+            with state_lock:
+                runs = RUNS + ([ORPHAN] if state["orphan"] else [])
+            self._json(200, {"runs": runs})
+            return True
+        if path == "/api/v1/projects":
+            self._json(200, {"projects": PROJECTS})
+            return True
+        if path == "/api/v1/repos":
+            self._json(200, {"repos": []})
+            return True
+        parts = path.split("/")
+        # /api/v1/projects/<id>/members
+        if len(parts) == 6 and parts[3] == "projects" and parts[5] == "members":
+            pid = urllib.parse.unquote(parts[4])
+            self._json(200, {"members": [
+                {"id": f"{pid}:crew.run:{ref}", "project_id": pid, "member_kind": "crew.run",
+                 "member_ref": ref, "meta": None, "attached_at": 1, "attached_by": "studio"}
+                for ref in MEMBERS.get(pid, [])
+            ]})
+            return True
+        # /api/v1/projects/<id>/interactive/api/docs — the notes project's registry
+        if path.startswith("/api/v1/projects/") and path.endswith("/interactive/api/docs"):
+            pid = urllib.parse.unquote(path.split("/")[4])
+            self._json(200, NOTES_DOCS if pid == "notes" else [])
+            return True
+        # /api/v1/runs/<id>/gate
+        if len(parts) == 6 and parts[3] == "runs" and parts[5] == "gate":
+            rid = urllib.parse.unquote(parts[4])
+            if rid == "r-q3":
+                with state_lock:
+                    age = state["q3_gate_age_ms"]
+                self._json(200, {"runId": rid, "ord": 0, "lifecycle": "open",
+                                 "prompt": "Approve the deck outline?",
+                                 "receivedAt": iso(NOW0 - age)})
+            elif rid == "r-api":
+                # `options: null` = free text ⇒ the COMPLEX gate shape (§7.11).
+                self._json(200, {"runId": rid, "ord": 0, "lifecycle": "open",
+                                 "prompt": "How should the tables move?",
+                                 "receivedAt": iso(NOW0 - 2 * MIN), "options": None})
+            else:
+                self._json(404, {"error": f"no gate cached for {rid}"})
+            return True
+        # /api/v1/runs/<id>/events
+        if len(parts) == 6 and parts[3] == "runs" and parts[5] == "events":
+            rid = urllib.parse.unquote(parts[4])
+            self._json(200, {"events": RUN_EVENTS.get(rid, [])})
+            return True
+        if path.startswith("/api/v1/"):
+            self._json(404, {"error": f"w2 fixture: no such endpoint {path}"})
+            return True
+        return False
+
+    def do_GET(self):  # noqa: N802 (stdlib naming)
+        if self.headers.get("Upgrade", "").lower() == "websocket":
+            return self._ws()
+        path = urllib.parse.urlparse(self.path).path
+        if self._api(path):
+            return None
+        if not Path(self.translate_path(self.path)).is_file():
+            self.path = "/index.html"  # client-side routes resolve to the shell
+        return super().do_GET()
+
+    def do_POST(self):  # noqa: N802 (stdlib naming)
+        path = urllib.parse.urlparse(self.path).path
+        if path == "/__fixture":
+            body = json.loads(self.rfile.read(int(self.headers.get("Content-Length") or 0)) or b"{}")
+            with state_lock:
+                state.update({k: v for k, v in body.items() if k in state})
+                snapshot = dict(state)
+            return self._json(200, {"ok": True, "state": snapshot})
+        return self._json(404, {"error": f"w2 fixture: no such endpoint {path}"})
+
+
+def ensure_build(fail) -> Path:
+    """The same-origin build (shared across the rigs — same dist dir).
+    `fail(step, why)` is the calling rig's reporter; SKIP_STUDIO_BUILD=1 skips."""
+    dist = REPO / "dist-sameorigin"
+    if os.environ.get("SKIP_STUDIO_BUILD") == "1":
+        if not (dist / "index.html").is_file():
+            fail("build", f"SKIP_STUDIO_BUILD=1 but {dist}/index.html is missing — "
+                 "build it with `npx vite build --outDir dist-sameorigin` (no VITE_API_HOST)")
+    elif not (dist / "index.html").is_file():
+        env = dict(os.environ, VITE_API_HOST="")
+        r = subprocess.run(
+            [NPM, "exec", "--", "vite", "build", "--outDir", "dist-sameorigin", "--emptyOutDir"],
+            cwd=REPO, env=env, capture_output=True, text=True, timeout=600,
+        )
+        if r.returncode != 0:
+            fail("build", f"same-origin vite build failed:\n{r.stdout[-2000:]}\n{r.stderr[-2000:]}")
+    return dist
+
+
+def start_server(port: int, dist: Path) -> str:
+    """Serve `dist` + the fixture API on 127.0.0.1:`port` (daemon thread); returns the origin."""
+    httpd = ThreadingHTTPServer(("127.0.0.1", port), partial(W2Handler, directory=str(dist)))
+    threading.Thread(target=httpd.serve_forever, daemon=True).start()
+    return f"http://127.0.0.1:{port}"
+
+
+def set_fixture(origin: str, **kwargs) -> None:
+    """Flip the mutable switches over POST /__fixture between page loads."""
+    req = urllib.request.Request(f"{origin}/__fixture", method="POST",
+                                 data=json.dumps(kwargs).encode())
+    req.add_header("Content-Type", "application/json")
+    with urllib.request.urlopen(req, timeout=10) as res:
+        res.read()
